Create Boogle.java
wrote down pseudocode-ish thought process because I have no idea how to actually code. I originally had a misunderstanding that binary search was cutting(?) the arraylist in half and THEN traverse (which is linear search) to make the computer process it faster, but that wasn't the case. (this is probably reflected in the pseudocode.) 

Also, I learned that binary search can only work in a sorted list. Topher suggested (after class zoom session) that we can add isSorted() to see if the list is actually sorted (just check if the next array element is bigger than the one that you're in right now).

I am still struggling to understand certain parts, like why do we have to copy the array list first or make a temporary node, but this is probably more conceptual thing that I am trying to understand and need some time to let that sink in....

Onward.