Create BSTree.java
the if statements must not be complete, because it keeps giving me NullPointerException. I need to trace it so that I can point out exactly why it's wrong, but it made total sense when I wrote the pseudocode lol I put System.out.println so that I can see what's the current data after if statement, and it seems like currentData is not being updated! Do I have to do something like current.getRight().getData()???